+++ pages/list.tsx
@@ -1,8 +1,8 @@
 import { Noto_Sans } from 'next/font/google'
 
-import FilteringField from '@/features/common/components/FilteringField'
 import Navbar from '@/features/common/components/Navbar'
 import { classNames } from '@/features/common/utils/classNames'
+import FilteringField from '@/features/products/components/FilteringField'
 import Item from '@/features/products/components/Item'
 
 const NotoSans = Noto_Sans({
